Add GraphDataset superclass
