+++ data_prep/agnews_graph.py
@@ -3,11 +3,11 @@ import torch
 from torch_geometric.data import Data
 
 from data_prep.graph_utils import pmi, tf_idf_mtx
-from data_prep.dataset import Dataset
+from data_prep.dataset import GraphDataset
 from datasets import load_dataset
 
 
-class AGNewsGraph(Dataset):
+class AGNewsGraph(GraphDataset):
     def __init__(self, device, val_size=0.1):
         """
         Creates the train, test, and validation splits for AGNews.
@@ -69,68 +69,6 @@ class AGNewsGraph(Dataset):
         self.data.train_mask = train_mask
         self.data.val_mask = val_mask
         self.data.test_mask = test_mask
-
-    def generate_edges(self, num_docs, tf_idf, pmi_scores):
-        """Generates edge list and weights based on tf.idf and PMI.
-
-        Args:
-            num_docs (int): Number of all documents in the dataset
-            tf_idf (SparseMatrix): sklearn Sparse matrix object containing tf.idf values
-            pmi_scores (dict): Dictonary of word pairs and corresponding PMI scores
-
-        Returns:
-            edge_index (Tensor): List of edges.
-            edge_attr (Tensor): List of edge weights.
-        """
-        edge_index = []
-        edge_attr = []
-
-        # Document-word edges
-        for d_ind, doc in enumerate(tf_idf):
-            word_inds = doc.indices
-            for w_ind in word_inds:
-                edge_index.append([d_ind, num_docs + w_ind])
-                edge_index.append([num_docs + w_ind, d_ind])
-                edge_attr.append(tf_idf[d_ind, w_ind])
-                edge_attr.append(tf_idf[d_ind, w_ind])
-
-        # Word-word edges
-        for (word_i, word_j), score in pmi_scores.items():
-            w_i_ind = self.ntoi[word_i]
-            w_j_ind = self.ntoi[word_j]
-            edge_index.append([w_i_ind, w_j_ind])
-            edge_index.append([w_j_ind, w_i_ind])
-            edge_attr.append(score)
-            edge_attr.append(score)
-
-        edge_index = torch.tensor(edge_index).t().contiguous()
-        edge_attr = torch.tensor(edge_attr).float()
-        return edge_index, edge_attr
-
-    def generate_masks(self, train_num, val_num, test_num):
-        """Generates masking for the different splits in the dataset.
-
-        Args:
-            train_num (int): Number of training documents.
-            val_num (int): Number of validation documents.
-            test_num (int): Number of test documents.
-
-        Returns:
-            train_mask (List): Training mask as boolean list.
-            val_mask (List): Validation mask as boolean list.
-            test_mask (List): Test mask as boolean list.
-        """
-        val_mask = torch.zeros(len(self.iton), device=self.device)
-        val_mask[:val_num] = 1
-
-        train_mask = torch.zeros(len(self.iton), device=self.device)
-        train_mask[val_num:val_num + train_num] = 1
-
-        # Mask all non-test docs
-        test_mask = torch.zeros(len(self.iton), device=self.device)
-        test_mask[val_num + train_num:val_num + train_num + test_num] = 1
-
-        return train_mask.bool(), val_mask.bool(), test_mask.bool()
 
     @staticmethod
     def prepare_agnews(val_size=0.1):
